CTABLES: Fix treatment of string variables that differ past the first byte.
CTABLES display was treating string values with more than one byte as the
same if they had the same first byte, because var_get_type() was confused
with var_get_width().  (The former returns 0 for numeric and 1 for string
and the latter returns 0 for numeric and string width for string.)

Bug #64044.