combine-files: Adopt the encoding of the first input file.
Previously, there was no good way to control the output encoding, which was
simply the default encoding.  One could use SET LOCALE, but it seems more
reasonable to use the encoding from one of the input files, and the first
one is the most obvious choice.

Also, improve the warning about incompatible input encodings.