pspp-output: Use page setup from command line by default.
pspp-output was ignoring the page setup specified on the command line in
favor of the one embedded in an SPV file.  This is usually wrong, and in
particular it screwed up the examples in the manual by causing them to be
broken across multiple PNG output files, only the first of which was
actually shown in the manual.