gui: Update how the Run menu works in the syntax window.
Until now, the Run menu had both "Selection", which ran the current
selection, and "Current line", which ran the current line.  This commit
replaces that by a single menu item "Selection", which runs the current
command if there is no selection and all the selected or partially selected
commands if there is a selection.  This behavior matches SPSS.

Requested by Matthias Faeth.