MATCH FILES: Fix crash in string variable resizing.
Commmit 638bf380e550 ("combine-files: Allow matching string
variables to have different widths.") introduced a bug in the
case where the variables in data files being combined didn't
have the same variable order.  This is kind of an
embarrassing bug since the input files having the same order
of variables is an exceptional case anyhow.  This commit
should fix the problem.

The test as updated will crash if the bug regresses.

Thanks to Frans Houweling for reporting the bug.