tests: Stop using squish-pty.
squish-pty seems to break easily, so switch to using another way to
simulate interactive behavior.  This way requires a small source code
change to PSPP itself but it should be more portable.

Thanks to Friedrich Beckmann for reporting the problem.  The problem that
Friedrich reported was a hang in the "interactive output appears
immediately" test on Debian sid.  I couldn't reproduce that problem on
Fedora or Debian testing, but I did get failures on Debian testing.  I
guess it's best to just avoid ptys at all.