tests: Remove two tests that proved problematic cross-OS and cross-arch.
The test "journal enabled by default interactively" failed in Debian
testing on s390x, powerpc, and ppc64.  I can't reproduce the problem on
x86_64.  This commit removes it.

The test "interactive output appears immediately" failed in Debian testing
on s390x, hppa, powerpc, and ppc64.  It also previously failed on Mac OS.
I can reproduce the problem on x86_64 if readline isn't available, but not
if it is.  (I think it would always be available for Debian).  This commit
removes it.

Thanks to Friedrich Beckmann for reporting the Debian build problems.