rust: Major additions.
This adds support for writing system files to the Rust pspp library and
to the `convert` command in the Rust CLI.  It also renames the `dissect`
command in the CLI to `show` and adds numerous features to it.

This makes many bug fixes across the Rust pspp library and changes the
library internals in several ways.