gui: handle negative selections (bug #67215)
When you select cases starting from a higher case number to
a lower case number, then the selection start_y is higher than
the end_y. This results in a negative number of cases that is
supposed to be deleted. That resulted in a crash. This fix
handles this situation sucht that always a positive number of
cases is computed, independet from the way the selection is done.

Closes: https://savannah.gnu.org/bugs/?67215