fix corrupted odt file format (bug #67208)
The odt file that pspp wrote was identified as corrupted by
libreoffice. This was due to a possible mismatch between the
local header and the central directory entry for the bitflags.

This patch makes the local and central setting for the bitflags
consistent. Libreoffice now reads the odt without noise.

Thanks to Elias who reported the problem!

Closes: https://savannah.gnu.org/bugs/index.php?67208